Refactor conversation slice
