+++ .vscode/settings.json
@@ -1,11 +1,12 @@
 {
   "cSpell.words": [
+    "Chatwoot",
     "linkify",
     "markdownit",
     "middlewares",
     "persistor",
     "Pressable",
-    "Swipeable",
-    "Chatwoot"
+    "reduxjs",
+    "Swipeable"
   ]
 }
